% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coldata}
\alias{coldata}
\title{coldata file}
\format{
A data frame with rows corresponding to the number of samples and 6
variables
}
\description{
Data frame that describes samples composing the experiment.
}
\details{
\itemize{
\item Name.
\item condition. Condition of the sample, e.g. treatment or time points in case of time courses
\item type.
\item code.
\item replicates. ID of the replicate
\item cellLine. Cell line the sample belong to. In place of cell line, there can be the patient number
}
}
\keyword{coldata}
