% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIEP-package.R
\docType{package}
\name{MIEP-package}
\alias{MIEP-package}
\title{MIEP: Gene set-focused analysis of RNA-seq data}
\description{
Make-it-easy-pipeline (MIEP) is an integrated, interactive, and user-friendly R package that aggregates a high number of tools for the analysis of RNA-seq data. A modern approach to biological targets identification is implemented in MIEP. Since pathway disruption plays a critical role in cancer development by affecting the signaling mechanisms that rule cell function, focusing on gene sets mimicking pathways can accelerate the development and testing of treatments acting on key elements of signal transduction. MIEP pipeline develops and edits new gene sets based on the results of multiple integrated analyses.
}
\author{
\strong{Maintainer}: Alberto Corradin \email{alberto.corradin@iov.veneto.it}

Authors:
\itemize{
  \item Vincenzo Ciminale
}

}
\keyword{internal}
