% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIEP-package.R, R/MIEP.R
\name{MIEP}
\alias{MIEP}
\title{MIEP function launches the Make-it-easy-pipeline for gene set-focused
analysis of RNA-seq data}
\usage{
MIEP(filePath1 = NULL, filePath2 = NULL)
}
\arguments{
\item{filePath1}{Path to the pipeline folder. Inside it, data folder should
be located in order to allow subsequent uploading of data files (included
raw_gene_counts.tsv, coldata.csv, controls.txt,infoExprSet.tsv, and
genesetsLists.txt). If MIEP is called without passing this parameter's value
directly, the same information on the path will be asked by MIEP
interactively by waiting the selection of a file to be uploaded. First line
of such a text file should report the path to pipeline folder, and will
correspond to parameter filePath1.}

\item{filePath2}{Path to the folder containing Gene Matrix Transposed files
(*.gmt) from GSEA web site (www.gsea-msigdb.org/gsea/index.jsp), which
define composition of gene sets of interest.}
}
\value{
A folder with the results of the pipeline. MIEP performs statistical
testing, annotates sequences with gene ID names based on HGNC nomenclature
(https://www.genenames.org/), corrects potential biases due to cell lines
specificity (or batch effect), summarizes the main results by means of
tables in HTML format, heat maps, and dendrograms based on hierarchical
clustering. The latter can be customized to better highlight differences by
specifically designed color palettes. Results' folder will be automatically
created by MIEP inside the pipeline folder. Different results can be
obtained by inserting different parameters interactively or by selecting
different filtering process. If gene sets are supplied MIEP will provide
heat maps focused on gene sets, separately for each of them, volcano plots,
and tables showing results of statistical tests in HTML format.
}
\description{
MIEP function launches the Make-it-easy-pipeline for gene set-focused
analysis of RNA-seq data
}
\examples{
MIEP(filePath1=NULL,filePath2=NULL);
MIEP();
}
