% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw_gene_counts}
\alias{raw_gene_counts}
\title{Raw gene counts}
\format{
A data frame with rows corresponding to measured features and columns
to samples
}
\description{
Data frame containing the reads' copy numbers. This data was created in
silico, starting from data downloaded from  GEO web site
(https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE198518). 3 in silico
replicates for every sample in the original data set. Original data were not
inserted in this in silico data set.
}
\details{
\itemize{
\item GeneID. Sequence's Ensemble ID
\item Patient1_TREATED_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient1_TREATED_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient1_TREATED_rep3. Raw gene counts of patient 1, treated sample, replicates 3
\item Patient1_CNTR_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient1_CNTR_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient1_CNTR_rep3. Raw gene counts of patient 1, treated sample, replicates 3
\item Patient2_TREATED_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient2_TREATED_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient2_TREATED_rep3. Raw gene counts of patient 1, treated sample, replicates 3
\item Patient2_CNTR_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient2_CNTR_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient2_CNTR_rep3. Raw gene counts of patient 1, treated sample, replicates 3
\item Patient3_TREATED_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient3_TREATED_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient3_TREATED_rep3. Raw gene counts of patient 1, treated sample, replicates 3
\item Patient3_CNTR_rep1. Raw gene counts of patient 1, treated sample, replicates 1
\item Patient3_CNTR_rep2. Raw gene counts of patient 1, treated sample, replicates 2
\item Patient3_CNTR_rep3. Raw gene counts of patient 1, treated sample, replicates 3
}
}
\keyword{raw_gene_counts}
